requirements.txt.txt
Flask==2.3.2
mysql-connector-python==8.0.33
numpy==1.24.3
pygame==2.2.0
python-dotenv==1.0.0
gunicorn==20.1.0